requirements.txt
fastapi
uvicorn 
google-cloud-bigquery 
python-multipart
pydantic[email]
google-cloud-aiplatform
google-auth
matplotlib
seaborn
pandas
vertexai
datetime
pydantic
db-dtypes
google-cloud-storage
scipy
langchain_google_genai
smogn